refactor: separate deletion handlers from utility functions
